# backend-foryou/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "ForYou API Documentation",
        "description": "API documentation for the ForYou project",
        "contact": {
            "email": "admin@foryou.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/detalles-solicitud": {
            "get": {
                "tags": [
                    "Detalles Solicitud"
                ],
                "summary": "Get list of detalle solicitudes",
                "description": "Returns list of detalle solicitudes",
                "operationId": "getDetallesSolicitud",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/DetalleSolicitud"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Detalles Solicitud"
                ],
                "summary": "Create a new detalle solicitud",
                "description": "Creates a new detalle solicitud",
                "operationId": "createDetalleSolicitud",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DetalleSolicitudInput"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Detalle Solicitud created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DetalleSolicitud"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
        "/api/detalles-solicitud/{id}": {
            "get": {
                "tags": [
                    "Detalles Solicitud"
                ],
                "summary": "Get detalle solicitud by ID",
                "description": "Returns a single detalle solicitud",
                "operationId": "getDetalleSolicitudById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of detalle solicitud to return",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DetalleSolicitud"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Detalle Solicitud not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Detalles Solicitud"
                ],
                "summary": "Update an existing detalle solicitud",
                "description": "Updates a detalle solicitud",
                "operationId": "updateDetalleSolicitud",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of detalle solicitud to update",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/DetalleSolicitudInput"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Detalle Solicitud updated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/DetalleSolicitud"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Detalle Solicitud not found"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Detalles Solicitud"
                ],
                "summary": "Delete a detalle solicitud",
                "description": "Deletes a single detalle solicitud",
                "operationId": "deleteDetalleSolicitud",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of detalle solicitud to delete",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Detalle Solicitud not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "DetalleSolicitud": {
                "title": "DetalleSolicitud",
                "required": [
                    "id_solicitud"
                ],
                "properties": {
                    "id_detalles": {
                        "description": "ID del detalle de solicitud",
                        "type": "integer",
                        "format": "int64"
                    },
                    "id_solicitud": {
                        "description": "ID de la solicitud a la que pertenece este detalle",
                        "type": "integer",
                        "format": "int64"
                    },
                    "med_alt": {
                        "description": "Medida de alto",
                        "type": "number",
                        "format": "float"
                    },
                    "med_anc": {
                        "description": "Medida de ancho",
                        "type": "number",
                        "format": "float"
                    },
                    "descripcion": {
                        "description": "Descripción del elemento",
                        "type": "string"
                    },
                    "created_at": {
                        "description": "Fecha de creación",
                        "type": "string",
                        "format": "date-time"
                    },
                    "updated_at": {
                        "description": "Fecha de actualización",
                        "type": "string",
                        "format": "date-time"
                    }
                },
                "type": "object"
            },
            "DetalleSolicitudInput": {
                "title": "DetalleSolicitud Input",
                "required": [
                    "id_solicitud"
                ],
                "properties": {
                    "id_solicitud": {
                        "description": "ID de la solicitud a la que pertenece este detalle",
                        "type": "integer",
                        "format": "int64"
                    },
                    "med_alt": {
                        "description": "Medida de alto",
                        "type": "number",
                        "format": "float"
                    },
                    "med_anc": {
                        "description": "Medida de ancho",
                        "type": "number",
                        "format": "float"
                    },
                    "descripcion": {
                        "description": "Descripción del elemento",
                        "type": "string"
                    }
                },
                "type": "object"
            },
...
        }
    },
...
}